extract private class
